move SubscriptionCard story and component to Component dir from Views dir
