+++ src/stories/Components/SubscriptionCard.js
@@ -2,6 +2,6 @@ import React from "react";
 
 import { storiesOf } from "@storybook/react";
 
-import SubscriptionCard from "../../Views/SubscriptionCard";
+import SubscriptionCard from "../../Components/SubscriptionCard";
 
 storiesOf("Subscription card", module).add("cards", () => <SubscriptionCard />);
